refactor(quadrics): extract raySphereHit to scaffold/ui/rayHit (#120 step 3)
The ray-sphere hit-test function body was triplicated across
Slider.ts, Preset.ts, and SectionTab.ts — identical bodies in all
three. SectionTab's own comment noted "a third tap-button-like
primitive would be the trigger for refactoring all three onto a
shared base"; the v0.6 scaffold extraction is that trigger.

Body extracted verbatim (no behavior change). The three callers
import via @/scaffold/ui/rayHit. Sets up step 7 to bundle the
three primitives' moves + grabRadiusMultiplier parameterization
without splitting the constant across files.
